Store Salesforce Program Ids on the Course model (#887)
* Add salesforce_program_id column to courses table

* Remove obsolete attr_accessor's from Course model

The Course model was originally a Program model and these accessors
were used to implement the original Sync From Salesforce (aka Sync to LMS).
Obsolete now. See this commit for where they came from and how they were used:
https://github.com/bebraven/platform/commit/0d41d9578f144236063dc187e92b84919252c881

* Fix specs. Program.Id must be 18 chars

* When Launching a Program, set the Id on the Course models

* Sync Program.Id's back to the Course models.

See comments for why this is helpful. It also makes it so that after
we deploy we just have to let the nightly sync go for all active Programs
to the corresponding Course models.

Also short-circuit the sync for missing Course models. Mostly to prevent
devs from messing each other up.