Show Zoom link in CLASS assignment (#890)
* Show the Zoom links for Learning Lab or Orientation.

Task: https://app.asana.com/0/1201131148207877/1200877445440509

* Add Home breadcrumps to attendance_events#new

* AttendanceEventsController specs

* AttendanceEventSubmissionsController specs

* Implement specs for AttendanceEventSubmissionAnswersController

* Easier to read AttendanceEvent#event_type_display

* Add null: false DB constraint to attendance_event#event_type

* Add old suffix for example Program.Id.

The ID must be 18 characters, but instead of introudicing a new suffix
ust leave the old one so the diff is less to look at in the code review.

* Fix specs

* Better error handling for staff.

Remove the ability for admin's to launch this page since it will fail
unless they happen to be an Enrolled Participant. Add a message talking
about why you can't see this page if you're not using an Enrolled Participant.
This will help with QA and development too if we have test users that
aren't setup properly.

In the future we could add an admin page that let's you choose a Participant
and view their Zoom link in order to troubleshoot issues, but it's a low
prioirty b/c the link should match what's in the Salesforce so just make
sure that one works.

* Fix specs!

* Fix specs!!

* Fix specs!!!

* Add validation for event_type

* require 'salesforce_api' in attendance_event_submission_answers_controller

Just in case it's the very first thing that is hit after a server restart.
Very unlikely but better to be consistent with how we require the lib files

* Be extra defensive with SFParticipant.send(:field_name)

Don't want to expose ourselves to accidentally getting something into
the database that would then end up displaying sensitive information
about a participant on the CLASS assignment. It can't happen right now
b/c of the hard-coded intermediate lookup, but it could be easy to mess
that defense up in a future refactor.

* Add DB check constraint for attendance_events#event_type

* Add validation for event_type as well.

* Code review feedback. Honeycomb field and non-optional event_type

* Fix specs

* Fix db:rollback

* Add language about needing to Masquerade as users.

This is only for admins if they try to view a CLASS assignment.