Smarter Zoom sync logic - regenerate links when thing change (#893)
* Add zoom_link_infos table.

In order to make the Zoom sync smarter, we need to detect when changes
happen to the registrant details that would require us to create / delete
the existing link. This table holds the information about existing links
required to determine what to do.

* Add helper methods to ZoomLinkInfo

- ZoomLinkInfo.parse
- ZoomLinkInfo#first_name_with_prefix
- ZoomLinkInfo#registrant_details_match?

* SFParticipant#add_to_honeycomb_span()

When adding SFParticipant info to Honeycomb, let's do it consistently
using a method that takes an instance of the struct.
Apparently struct's can take a block which is a way to define methods
that act on instances of the struct. Cool!

* Add CURRENT_TIMESTAMP default to make upserts easier.

Note that we have some old migrations that use the pre-Ruby 3 syntax
which would fail if you ran the full migrations. This is a good reference for
how to fix those. Just change the param from a hash to a keyword param when
calling `t.timestamps(default: -> { "CURRENT_TIMESTAMP" })`

* Change ZoomAPI#add_registrant to take params instead of a hash.

External consumers shouldn't need to know the details of what the
API expects. It's cleaner not to have to create the hash to pass
in when doing the sync (and will make the .csv parsing a bit uglier
but we should be doing better parsing / error handling there anyway)

* Reimplement sync_zoom_links_for_participant to be smarter.

Instead of creating Zoom links only if they are empty or if forced,
detect if the information using to create them changed and update
them if necessary.

Note: this could be even smarter in a future pass and delete/cancel
the old registration if we create a new one. This happens only if
the email or meeting_id changes. A new registrant is created but the
old one is left hanging around. If only the first or last name change,
the link generated is new but it's the same registrant returned.

* Cutover ZoomAPI#add_registrant to take explicit params.

It used to take a hash and rely on the consumer to know what
should be in the hash. Now it takes the values explicitly and
builds the hash that the ZoomAPI expects.

* Handle "429 Too Many Requests" error

* Add TODO for cancelling old links.

Also, move the set_zoom_link_infos to a better spot in the file
so it's easier to read the service from top to bottom.

* Bug fix.

* Update "Force Zoom links" language in UI

* Fix users_roles_controller_spec

* Fix sync_zoom_links_for_participant

Fix existing specs and add a TODO to actually implement specs for the new sync logic.

* Refactor for consistency.

* Clarify instructions.

* Add tests for saving the ZoomLinkInfo back to the database

* Implement specs for new smarter sync logic.

* Replace NOOP with actual implementation.

Even though the Mock:SalesforceParticipant struct is a hacky thing
no reason to have the add_to_honeycomb_span() method be a NOOP on it.

* More consistent style with codebase.

We generally use && vs `and`

* Honeycomb fields

* Add get_registrants to ZoomAPI in case we need it for scripting.

It's currently not used in the code, but if I need it as part of running
a script when deploying it will be convenient to have there.

* Add check_constraint on Salesforce Participant ID length.

It must be 18 chars. We use upsert which skips validations so having
a DB level constraints is useful.