Heroku Connect: read-only access to Salesforce data with Rails models (#938)
* First pass at Heroku Connect support.

See Salesforce::SalesforceRecord for a big write-up of what that is.

* Sample postgres connection URL for test env breaks CI

Works find in local rspec? Cross this bridge once we try
to actually write specs

* Reduce selected columns to what we care about.

This is unfortunate, b/c I don't want to have to maintain a list
but I looked at the Contact table in prod and it has 95 columns. Yikes.
This is b/c we add any and everything we need in Periscope.

I think it's fine to maintain the list of fields we care about since
everytime we add a new one in Salesforce, we have to add it to the Mapping
in Heroku Connect, and it's just a small skip and a jump to add it to the
list in the model. The key here will be to make sure that folks can easily
figure out how to add a SF field and make it accessible in these models.
Maybe in a future pass we can throw an exception if a missing column/attribute
is tried with hints about where to look outside of the README / Wiki documentation
we should add to supplement the code comments.

* Add Salesforce::Cohort model for Cohort__c object

* Add Salesforce::CohortSchedule model for CohortSchedule__c object

* Add Salesforce::Candidate model for Candidate__c object

Also, define methods in the various models for things that are
formula fields and can't be synced with Heroku Connect. We'll
just implement the logic in platform code.

* Add Salesforce::Program model for Program__c object

Note: I'm realizing that the has_many relationships I've defined
don't work (for Contact too). Figure that out.

* Fix the has_many relationships.

Apparantly ActiveRecord didn't like using a symbol as the primary_key

* Clarify HEROKU_CONNECT_DATABASE_URL comments

* Remove TODO and add missing CohortSchedule#participants association

* Add TODOs around testing and secrets after discussing with engineers.

* Rename Salesforce namespace to HerokuConnect

We already use salesforce_xxx for a bunch of spec factories and stuff
so instead of dealing with renaming / reconciling that, since these
are Heroku Connect tables, just use that as the name. TODO: rename the
files themselves in a followup

* Rename files from salesforce/xxx to heroku_connect/xxx

* Typo

* Filter out isdeleted records.

Salesforce uses this status while something is in the process of being
deleted but isn't fully yet. We don't want those returned in normal
queuries.

* Add HerokuConnect::TaAssignment model for TA_Assignment__c object

* Implement teaching_assistant_sections logic for SFParticipant

* Implement zoom_prefix method for HerokuConnect::Cohort

Note: this handles empty LC names that shouldn't happen in prod but
can when setting up test users.

* Implement HerokuConnect::CohortSchedule#canvas_section_name

Mimic what the DayTime__c formula field does in SF. Heroku Connect
can't sync formula fields. Note that this also takes care of situations
that shouldn't happen in prod, but can happen when setting up test
schedules where no weekday or time is present.

* Remove obsolete TODO

* Remove signup_token and discord_invite_code fields.

They are secret and we should limit the places they are stored.
They are only supposed to be in Salesforce for the initial comms
but we shouldn't expose them in Platform or in Periscope to prevent
accidentally leaking them.

* The Student ID is personal info we don't need. Remove it.

We have a TODO to remove the places in the code where we used to use it
but this at least removes loading it into the Heroku Connect model

* Implement specs for HerokuConnect::Contact

* Implement specs for HerokuConnect::RecordType model

Note: this is the only one I've come across that doesn't have a
isdeleted column. Needed to change the base class to look for that
scenario.

* Implement specs for HerokuConnect::Program model

* Implement specs for HerokuConnect::CohortSchedule model

* Implement specs for HerokuConnect::Cohort model

* Add specs to check to indexs on HerokuConnect tables.

The point is to make sure that dev and prod match since it's easy
to create an index in the staging Sandbox but forget to add it in
prod before deploying.

* Implement specs for HerokuConnect::Candidate and HerokuConnect::Participant

* Implement specs for HerokuConnect::TaAssignment model

* Add TODO for missing Candidate specs

* Remove TODOs and add basic dev env setup to README

* Clarify some TODOs

* Implement specs for volunteer_role (renamed to candidate_role)

Also, change it to a symbol. We should be using symbols for things
like this, but haven't been consistent.

* Specs for HerokuConnect::CohortSchedule#canvas_section_name

* Implement HerokuConnect::Cohort#zoom_prefix specs

* Make sfid's unique per model in the factories

* Implement HerokuConnect::Participant#teaching_assistant_sections specs

* Remove TODO

* Clarify comment

* Clarify comments

* Add missing spec for HerokuConnect::Program#status__c

We query this for current and future, so it should have an index.

* Back out changes to SalesforceAPI constants.

In HerokuConnect, I want to be more consistent with string literal
vs symbol usage and had started to refactor the SalesforceAPI constants
to be symbols. On second thought, let's just make a clean break from
the SalesforceAPI for HerokuConnect stuff. We'll use symbols everywhere
where it makes sense but use new constants when using it. Eventually,
we'll refactor most of the SalesforeAPI to use HerokuConnect and
and tackle the features that use the constants as we refactor and
test

* Convert appropriate HerokuConnect attributes to symbols.

Also, add constants to compare known values.