Remove student_id from our code. (#947)
Task: https://app.asana.com/0/1201131148207877/1201453841518462
We used to use it as a fallback way to lookup
a user when we couldn't find them by email. Now we store both the Contact.ID
and User.canvas_user_id which are unique identifiers for a "person" and no
longer need to rely on student id. It adds a lot of complexity to try and get
that in the new Heroku Connect stuff (plus it may not exist or may be wrong)