SYNC: recurring scheduled jobs support + ParticipantSyncInfo model (#949)
* Add HerokuConnect::Program#current_and_future_program_ids scope

* Proof of concept for how to get the list of Participants to sync

* Add a ParticipantSyncInfo::Diff class to hold the old and new values.

We'll pass this to the sync services so they can easily tell what
parts of the info need to be synced and then mark the sync status
at the end so that it get's saved back at the end.

* Add contact_id to ParticipantSyncInfo and change role/status to symbols

contact_id is needed as part of the contact sync b/c that's how it
looks up the user.

* First pass of re-implementing the sync using ParticipantSyncInfos

Tons of TODOs and commented out code, but this works to sync the a
primary enrollment change.

* First pass at reimplementing TaAssignmnets sync with ParticipantSyncInfo

It works for the basic case, but not if you change an existing TaAssignment.
Need to fix.

* Fix TA Assignments sync.

We now store the TA Assignment related stuff with the actual
name of either the assignment section(s) or the ta caseload section
using SQL so we can detect the change to the names even if the IDs
of the TA Assignment stays the same

* Refactor calculated HerokuConnect fields to expose calculation to consumers.

We want to be able to run the calculation logic with a ParticipantSyncInfo
model's values without querying for the underlying HerokuConnect model,
so expose the calculation as a class method to share the logic

* Abstract out logic for calculated HerokuConnect functions.

It's still nice to have the values available on the models,
but we don't need to query the models from ParticipantSyncInfo
b/c we already have the values, so abstract it out and share the code

* Refactor Zoom link sync to use ParticipantSyncInfo instead of SalesforceAPI

* Remove obsolete TODO

* Remove obsolete TODO. Clarify comment.

* Add TODO

* Cleanup ParticipantSyncInfo.

Want to deploy this to prod and turn on a job to run it every X
minutes, but don't want to hook it up to the actual sync logic yet.
This gets ParticipantSyncInfo in a state where we can run it in prod
without risk

* Back out changes to the actual sync logic.

This PR will just have the new ParticipantSyncInfo.run_sync_poc
method running and saving the new values, but it won't actually run
any sync logic. Once that's run in prod long enough that we're
confident we don't have issues with "SSL SYSCALL error: EOF detected"
errors for the database connection, AND we've fully tested the new
refactored sync services that use ParticipantSyncInfo we'll deploy that
separately.

See here for the prototype:
https://github.com/bebraven/platform/pull/946

* Implement ParticipantSyncInfo.add_to_honeycomb_span()

* Add infrastructure for scheduling unique jobs.

These jobs act like cron jobs using the sidekiq-scheduler gem.
We can set their schedule in `config/sidekiq.yml`. They also have
the ability to control whether multiple jobs with the same perform()
parameters can run concurrently using the sidekiq-unique-jobs gem.

* Run a Proof Of Concept sync every 10 minutes (by default).

This sync just loops over all Participants, finds the changed ones,
sleeps for 2 seconds for each (to mimic a real sync), and saves the new
values back to the ParticipantSyncInfo model.

* Cleanup comments

* Update CronJob base class documentation.

* Move HerokuConnect::Participant#teaching_assistant_sections to ParticipantSyncInfo

This is a need for sync logic, not general HerokuConnect querying.
It belongs on the ParticipantSyncInfo instead.

* Link to Asana tasks for all TODOs

* Remove obsolete proof of concept code.

* Fix the lock_ttl to be the max expected run time of a sync.

* Move SyncSalesforceChangesJob and CronJob into a more sensical class hierarchy

Now we have: SyncSalesforceChangesJob < ScheduledJob < SidekiqJob

* Move all ActiveJobs to a sub-director of app/jobs

I thought it made more sense to have app/jobs/active_jobs and
app/jobs/sidekiq_jobs instead of putting them both at the root
level.

* Clarify comments around the sidekiq-unique-jobs options.

Hopefully makes it easier for the next dev to understand the decision
around the chosen options and edit them if necessary by linking to the
docs

* Make ActiveJob and SidekiqJob honeycomb instrumentation consistent

* Add ability to disable sidekiq-scheduler

Devs may not want it always running in their dev env

* Change participant_sync_infos#sfid to varchar(18) column type

We treat this as the primary key, so any small performance gain by setting
the length of this column precisly may be worth it.

Set the contact_id to this length as well while we're at it since it
also holds an 18 char Salesforce ID

* Bug fix.

enrollment_changed always returned false. No clue how this syntax
even worked

* Cleanup ENV vars.

Make SALESFORCE_SYNC_MAX_DURATION take absolute seconds instead of
eval'ing a string. We don't want to have eval()'s in the code whenever
possible b/c it's easy to forget to sanitize the input which is a
security concern.

The SALESFORCE_SYNC_EVERY option actually uses this gem for parsing
https://github.com/floraison/fugit so we can't use absolute seconds
to be consistent with SALESFORCE_SYNC_MAX_DURATION.

* Switch SALESFORCE_SYNC_EVERY to use seconds

Want to be consistent with SALESFORCE_SYNC_MAX_DURATION which takes
seconds to avoid calling eval() for security reasons (aka future refactors
or copy/pasta forgetting to sanitize the eval'd string).

Needed to put this in an array for it to work. Seems to be a bug in the
gem where if you use a string that can be directly converted to an integer,
it assume's you're using the Array format to specify a more advanced schedule.
E.g.
    every: ['30s', first_in: '120s']

* Cleanup comments