Hotfix: Honeycomb field was wrong for User failed auth. (#956)
In the cas_controller, if the auth fails it calls
`add_honeycomb_context(last_failed_authenticator.user)` so that we can
get info about the user failing to authenticate in Honeycomb. However,
this method was setting `@user` back to nil. With this change,
if a user is found for the email they are trying to login with, if auth
fails, then the Honeycomb fields for that user are still sent.