SYNC: remove "Send Signup Emails" feature from Program sync (#963)
* Fix "Send Signup Emails" feature.

We broke this feature in this PR: https://github.com/bebraven/platform/pull/922
It's only used when manually running the sync from the Platform Admin dashboard,
so default to false but pass it through the layers of the sync when it comes from
that entry point.

* When manually sending a Sign-Up email, redirect to users_path.

It was staying on the same page and you had to hit the back button 2
times to get back to the User (which has breadcrumbs to help navigate).

* Rip out the "Send Signup Emails" feature for the Program sync.

This is currently broken and wasn't working as intended in the first
place when it was, so just rip it out. Less work then retesting all
this and less to deal with in the sync refactor I'm working on here:
https://github.com/bebraven/platform/pull/950

Now, to send a signup email, either wait for Campaign Monitor
or go directly to this User and click the "Send Signup Email" button.

The feature was originally intended to have a staff member confirm
a Participant or a set of Participants and run the sync with this
setting on to send them an email if they were rushing to get them
sign-up links. However, there are so many more moving parts now that
this wasn't happening and it was very likely that a sync would have
already run (maybe by another prod support staff member) by the time
they were ready to run the sync with this setting on. It was just
causing confusion since it would only work the very first time.

* Missed 2 spots in the specs

* Fix specs