Allow NULL for ParticipantSyncInfo#cohort_schedule_weekday (#969)
When folks are Dropped, their Cohort Schedule is cleared out. Similarly,
before we know the University Section they are registered for we may not
have set their Cohort Schedule yet. It would be too hard / too much work
to enforce all Participants to have a Cohort Schedule in Salesforce, so
we need to allow it in Platform and make sure the sync logic handles it.

I'm thinking we change the sync logic to fail to sync their enrollment
if it's not set instead of doing what we currently do and put them in
a Default Section. The sync prior to creating their Canvas account should
not error. E.g. it should sync their Zoom links, their local Section,
etc. It's only once we send a sign-up link and they try to use it to create
a Canvas account that we don't know what to do if they don't have a schedule.

In Sentry, since I rolled out the proof of concept for usign ParticipantSyncInfos
to run the sync every 10 minutes, 2.5K PG::NotNullViolation errors:
https://sentry.io/organizations/braven-pr/issues/2873519086/?project=5639961&query=is%3Aunresolved