SYNC: Run the job once a minute and gracefully exit early if it hits the max duration (#972)
* Save each synced person + exit early if max time is hit.

When reviewing the PoC data in Honeycomb prod, I realized that there are
certain times when a lot of changes happen (like 400 Participants have a change
probably from a roster or other bulk upload). Even with the 2 second per changed
person sync, this exceeded the 15 minute SALESFORCE_SYNC_MAX_DURATION. Based on
prod benchmarking, the sync is more like 4 seconds (and up to 7s) per Participant
with changes. Bumping that in the PoC too.

This change is a step towards handling that:
1) exit out and stop processing folks if we're past the max duration
2) save each participant as they're synced.

This means that we can still leave the max duration short-ish (to handle
a job crashing or a server restart during a job) and increase the frequency
of the sync so that the next job will start soon and pick up where the previous
one left off.

* Change sync to run every minute by default.

Usually, this should be just a few seconds if nothing changed. If something did
change, it will now run for up to SALESFORCE_SYNC_MAX_DURATION and then
exit allowing the next sync to resume

* Add a custom "conflict: log_honeycomb" strategy for SidekiqUniqueJobs

Instead of just logging the fact that a job was skipped b/c another
one was running, also send that fact to Honeycomb so that I can query
for this in conjunction with the field sent when a job exits early
to better understand how this is behaving.

* Change SyncSalesforceChangesJob from 15 to 10 min runtime.

Now that the sync happens every minute and we gracefully exit if we
hit the max duration, we can make this a little shorter.

Also, switch the job to use the new Honeycomb conflict strategy.