Change "Braven LTI Extension" to "Braven" (#992)
Most pages with the lti_canvas layout set their own title,
but if the dev forgets to do that or they end up at the root
url with a "Not Authorized" exception, the browser title can
show "Braven LTI Extension" in the browser tab, which isn't especially
meaningful for end-users. Let's just be generic.