After running Launch Program a second time, clear out old local Courses. (#1034)
If Launch Program fails or we need to re-run it b/c we launched the wrong thing,
the local Courses will have both the old and new ones mapped to the Program. On
the next sync, this change makes it so that those get fixed up. This also applies
to if we manually change the Canvas Course IDs in Salesforce, usually for testing
purposes.