More Zoom link sync instrumentation. (#1036)
We ran into a situation during testing where the Zoom links where
set in Zoom but not in Salesforce. It was really hard to tell what
happeneding in Honecomb b/c it looked like we had sent them. I think
it's fine to put the actual links in HC. It's not a security concern.
It'll make it easier to see the Zoom sync changes and troubleshoot
this if if happens again.