HOTFIX: better error handling for Rise360Module Auto-grading (#1042)
- If a non-Student, aka a TA, goes through a whole Module, the auto-grading
  code will kick off when they hit 100% and lead to a 404 Not Found when
  trying to retrieve the submission
- If an Enrolled Fellow Participant is currently having issues where they are
  not being synced properly and we need to fix-up their Participant, like set
  a Cohort Schedule for example, the nightly grading task will completely fail
  once we hit that user.

This change handles the above two scenarios by skipping folks who we can't get
a submission for. It also adds course level error handling in case we have a test
course that we don't care about too much and isn't urgent to fixup, the rest of
auto-grading will still run even if things are throwing errors for that course.