Add calling class/method info to RestClient Honeycomb instrumentation (#1077)
* Add calling class/method info to RestClient Honeycomb instrumentation

We've had some noisy alerts for RestClient::NotFound that are expected
and handled in the code but we don't have a good way to tune our triggers
to ignore them. This adds not only the method being called in whatever
API wrapper class we're using (e.g. CanvasAPI), but also the parent
calling class and method of that API wrapper as well.

* Fix specs

* Be a little safer in how far down the stack we go

* Add some missing BOT ENV vars to example file.

* Fix DiscordBot instrumentation.

This is a hack just to make it more obvious that the RestClient call is
coming from the Discordrb gem code. In the bot, it doesn't really matter
much since it sends Honeycomb data to a separate dataset, but for the
Discord scheduler (or any future feature that runs in the application
code itself, it's less confusing to see "Discord" in the RestClient
class fields.