Upgrade to Rails 7.0 (#1090)
* Upgrade the least stuff possible to get the app to start.

There are errors and things are breaking, but the container doesnt die!

* Remove Spring.

Was getting this error:
You have already activated digest 3.0.0, but your Gemfile requires digest 3.1.0.

Looks like it's a Bootsnap / Spring weirdness. Rails 7 no longer ships with
Spring enabled by default, the spring-watcher-listen gem doesn't seem to be
maintained anymore and don't support v3 or v4 of Spring.

I removed it and tried things out. Error goes away. No noticable difference
from some quick sanity testing of Guard reloading changes or rake tasks
running.

* Migrate Keypair model from attr_encrypted gem to ActiveRecord encryption

Note: 3 new ENV vars are introduced that must be set in prod before deploying.
In all other environments they are hard-coded so devs don't have to manage them.

* Upgrade application defaults to Rails 7.

* Upgrade libhoney gem + add specs to Honeycomb traces

Spent awhile trying to upgrade the libhoney gem since it was
logging a TON of errors, so in the process of troubleshooting that
I went ahead and put each spec example in a trace of it's own.

* Remove open redirect that was mistakenly added.

Fix specs. The "target_uri" for the LtiLaunch is actually a platform
endpoint, i added the allow_other_host option accidentally and just had
to fix the spec