Performance tune Module grading for unregistered users. (#1108)
The motivation here was to tune out noisy alerts, but I think they
can still happen for other reasons, so I'm going to try and do that
in the trigger using the new restclient.parent_class and related fields.

However, in the period of time we've accepted a bunch of folks and they
start sycning and getting Canvas accounts, a failed CanvasAPI will happen
for each of them in the nightly grading. This change filters out folks
who haven't registered yet from nightly grading.